Remove unique interface not used elsewhere
